MESHFREE
- the bin walls and heat sources have been substituted with boost::unordered_map (hash_table) to improve performances.
- added SetUpWallDiscret in particle_algorithm to reuse the method where is needed
- some variable were renamed to improve consistency
- moved SetStateVector out of MF_density (for divfree dynamics)
- solved a bug with ITER_MAX (double2int)









git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22761 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_particle/particle_timint_impl.cpp
@@ -33,7 +33,7 @@ PARTICLE::TimIntImpl::TimIntImpl(
     output
   ),
   errToll_(particledynparams.get<double>("ERROR_TOLL")),
-  iterMax_(particledynparams.get<double>("ITER_MAX"))
+  iterMax_(particledynparams.get<int>("ITER_MAX"))
 {
   return;
 }
